Share syscall definitions
