+++ src/sgame/sg_trapcalls.h
@@ -40,7 +40,7 @@ int              trap_FS_Write( const void *buffer, int len, fileHandle_t f );
 void             trap_FS_Rename( const char *from, const char *to );
 void             trap_FS_FCloseFile( fileHandle_t f );
 int              trap_FS_GetFileList( const char *path, const char *extension, char *listbuf, int bufsize );
-void             trap_LocateGameData( gentity_t *gEnts, int numGEntities, int sizeofGEntity_t, playerState_t *clients, int sizeofGClient );
+void             trap_LocateGameData( int numGEntities, int sizeofGEntity_t, playerState_t *clients, int sizeofGClient );
 void             trap_DropClient( int clientNum, const char *reason );
 void             trap_SendServerCommand( int clientNum, const char *text );
 void             trap_SetConfigstring( int num, const char *string );
@@ -63,7 +63,6 @@ void             trap_SetUserinfo( int num, const char *buffer );
 void             trap_GetUserinfo( int num, char *buffer, int bufferSize );
 void             trap_GetServerinfo( char *buffer, int bufferSize );
 void             trap_AdjustAreaPortalState( gentity_t *ent, bool open );
-bool         trap_AreasConnected( int area1, int area2 );
 int              trap_BotAllocateClient();
 void             trap_BotFreeClient( int clientNum );
 void             trap_GetUsercmd( int clientNum, usercmd_t *cmd );
